requirements.txt
beautifulsoup4==4.11.1
discord.py==2.1.0
google_api_python_client==2.66.0
protobuf==4.21.9
python-dotenv==0.21.0
selenium==4.0.0a1
